+++ cordova-lib/spec-plugman/plugins/org.test.plugins.dummyplugin/src/windows/dummer.js
