+++ cordova-lib/spec-plugman/plugins/org.test.plugins.dummyplugin/src/windows/dummy3.vcxproj
