+++ cordova-lib/spec-plugman/plugins/org.test.plugins.faultyplugin/src/windows/faultyPlugin.js
