+++ cordova-lib/src/util/windows/jsprojManager.js
